requirements.txt
absl-py==0.8.1
annoy==1.16.2
astor==0.7.1
backcall==0.1.0
certifi==2019.9.11
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
gast==0.2.2
google-pasta==0.1.8
grpcio==1.23.0
h5py==2.10.0
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mkl-service==2.3.0
mock==3.0.5
mtcnn==0.1.0
numpy==1.17.3
olefile==0.46
opencv-python==4.1.1.26
opt-einsum==0+untagged.49.gdbede45.dirty
pandas==0.25.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
protobuf==3.10.1
Pygments==2.4.2
pygpu==0.7.6
pyparsing==2.4.5
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyreadline==2.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
scipy==1.3.1
six==1.13.0
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
Theano==1.0.4
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
